+++ CHANGELOG.rst
@@ -6,6 +6,15 @@ All notable changes to the of_lldp NApp will be documented in this file.
 [UNRELEASED] - Under development
 ********************************
 
+Changed
+=======
+- Subscribed to ``kytos/of_core.switch.interface(s).created`` to be more reliable when handling the of_lldp flow and respective vlan usage
+
+Fixed
+=====
+- Fixed potential out of order switch events when managing of_lldp flow
+
+
 [2025.2.0] - 2026-02-02
 ***********************